backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
requests>=2.31.0
pypdf>=5.0.0
python-multipart>=0.0.9
supabase>=2.10.0
